+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.2 branch
 ### 1.2.8
+* `can_be_string()` renamed as `is_stringable()`. `can_be_string()` is
+    deprecated and it will be removed in a later release;
 * added `is_absolute()` global function;
 * uses `symfony/filesystem` and `symfony/finder` packages.